完善

+++ src/main/kotlin/io/zhudy/duic/domain/PageResponse.kt
@@ -4,6 +4,6 @@ package io.zhudy.duic.domain
  * @author Kevin Zou (kevinz@weghst.com)
  */
 data class PageResponse(
-        val total: Int = 0,
+        val total: Long = 0,
         val items: Any = emptyList<Any>()
 )
